% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{ari}
\alias{ari}
\title{Calculate adjusted rand index}
\usage{
ari(x, y)
}
\arguments{
\item{x, y}{Vector of classifications}
}
\description{
This function is taken verbatim from the \code{mclust} package.
}
