% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.by}
\alias{as.data.frame.by}
\title{Coerce objects of class "by" to a data.frame}
\usage{
\method{as.data.frame}{by}(x, row_suffix = F, ...)
}
\arguments{
\item{x}{An object of class "by".}

\item{...}{Other arguments passed to data.frame}
}
\value{
A data.frame
}
\description{
Coerce objects of class "by" to a data.frame
}
\details{
Columns are converted to numeric where as.numeric does not produce any
additional NA values.
}
