% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.infl}
\alias{as.data.frame.infl}
\title{Coerce objects of class "infl" to a data.frame}
\usage{
\method{as.data.frame}{infl}(x, ...)
}
\arguments{
\item{x}{An object of class "infl" (produced by \code{\link{influence.measures}} or \code{\link{influence_measures}}.}

\item{...}{Other arguments passed to data.frame}
}
\value{
A data.frame
}
\description{
Coerce objects of class "infl" to a data.frame
}
