% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canola}
\alias{canola}
\title{Measurements of canola (\emph{Brassica napus} L.) height}
\format{
A data frame containing 18 measurements:
\describe{
\item{\strong{height_cm}:}{Plant height, in cm}
}
}
\source{
Cole, B. (1998). The use of morphological and RAPD markers for distinguishing
canola (Brassica napus L.) varieties. [Master’s thesis, University of Guelph]. UG
Atrium Repository. \url{https://hdl.handle.net/10214/20400}.
}
\usage{
canola
}
\description{
Measurements of canola (\emph{Brassica napus} L.) height
}
\keyword{datasets}
