% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{collinearity_diagnostic}
\alias{collinearity_diagnostic}
\title{Diagnosing multicollinearity in regression models}
\usage{
collinearity_diagnostic(object, digits = 4)
}
\arguments{
\item{object}{A fitted object.}

\item{digits}{the number of digits to print}
}
\description{
Calculate eigenvalues, conditional index and variance decomposition proportion
for a fitted model.
}
\references{
Hallahan, Charlie. 1995. “Understanding the Collineartiy Diagnostics in
Sas/Insight and Proc Reg.” In NESUG 1995, 801–4. Washington, D.C: Northeast SAS Users
Group. https://lexjansen.com/nesug/nesug95/NESUG95147.pdf.
}
