% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompanion.R
\name{Cramer_v}
\alias{Cramer_v}
\title{Cramer's V Test}
\usage{
Cramer_v(
  x,
  y = NULL,
  ci = FALSE,
  conf = 0.95,
  type = "perc",
  R = 1000,
  histogram = FALSE,
  digits = 4,
  bias.correct = FALSE,
  reportIncomplete = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{}
}
\description{
Cramer's V Test
}
\references{
Mangiafico, S. (2021). "rcompanion: Functions to Support
Extension Education Program Evaluation". R package version 2.3.27.
\url{https://CRAN.R-project.org/package=rcompanion}
}
