% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{cv_confint}
\alias{cv_confint}
\title{Confidence intervals for coefficient of variation}
\usage{
cv_confint(mean, sd, n, alpha = 0.05, pct = T)
}
\arguments{
\item{mean}{Numeric. A mean.}

\item{sd}{Numeric. A standard deviation.}

\item{n}{Integer. Number of observations.}

\item{alpha}{Numeric. Type I error rate for confidence intervals. Defaults
to 0.05.}

\item{pct}{Logical. If TRUE (the default), results are expressed as percents.}
}
\value{
An array, containing the coefficient of variation, and the upper and
lower confidence intervals for the given alpha.
}
\description{
This function calculates the coefficient of variation and
confidence intervals for a fitted model using the modified MacKey method,
as described by Vangal (DOI: 10.1080/00031305.1996.10473537, see References
section for complete citation). The code is adapted from the SAS macro
cvmixed, by Stephen Bowley
}
\references{
Vangel MG (1996). "Confidence Intervals for a Normal Coefficient
of Variation". \emph{The American Statistician}, \emph{50}(1), 20-26. doi:
\href{https://doi.org/10.1080/00031305.1996.10473537}{10.1080/00031305.1996.10473537}
}
