% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{dffits}
\alias{dffits}
\alias{influence_measures}
\alias{influence.lme}
\title{Additional methods for calculating influence measures}
\usage{
dffits(model, ...)
}
\arguments{
\item{model}{An fitted model}

\item{...}{Additional arguments passed to other functions.}

\item{groups}{Character. An optional column name in \code{data} defining the groupings to be removed in calculating leave-one-out deletion diagnostics. If not supplied, defaults to each row in \code{data}.}

\item{data}{A data.frame. The data used in fitting the original model. Only required when no method for \code{getData} exists.}

\item{infl}{An object created by calling \code{influence()}.}

\item{ncores}{Integer. The number of cores to use for parallel computing of influence measures.}

\item{level}{Integer. Level of conditioning in calculating residuals, predictions, etc. Level = 0 returns marginal residuals.}
}
\description{
Additional measures of influence and case-deletion diagnostics for regression models.
}
\details{
When the HGSB package is loaded, S3 methods for models of class \code{lm},
\code{glm} and \code{mlm} are registered which use the code from the
\code{stats::covratio}, \code{stats::dffits} functions. These base functions are then
masked by generics of the same name, allowing these methods to be extended to
additional classes of models.

The following methods are then registered for models of class \code{lme}:
\code{influence}, \code{dfbeta} and \code{dfbetas} (code adopted/adapted from Fox and
Weisberg, 2019); \code{hatvalues}, \code{cooks_distance} and \code{dffits} (following
the approaches described in Christensen, Pearson & Johnson, 1992), and
\code{covratio}, \code{rstandard}, and \code{press} (following the methods described
in SAS Institute, 2011). Note that, somewhat confusingly, \code{rstandard} calculates
what SAS refers to as internally studentized residuals, not what SAS defines as
standardized residuals.

In the future, these methods may be extended to models fit
using other functions from the \code{nlme} package (e.g. \code{nlme::gls}), or other
packages entirely (e.g. \code{glmmTMB}).
}
\references{
Christensen, Ronald, Larry M. Pearson, and Wesley Johnson. 1992.
“Case-Deletion Diagnostics for Mixed Models.” Technometrics 34 (1): 38.
URL:https://doi.org/10.2307/1269550.

John Fox and Sanford Weisberg. 2019. An {R} Companion to Applied
Regression, Third Edition. Thousand Oaks CA: Sage.
URL:https://socialsciences.mcmaster.ca/jfox/Books/Companion/

SAS Institute Inc. 2011. "The Mixed Procedure: Residuals and Influence
Diagnostics" in the SAS/STAT® 9.3 User’s Guide. Cary, NC:SAS Institute Inc.
}
