% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdROC.R
\name{diagnostic_summary}
\alias{diagnostic_summary}
\title{Summarize the performance of a diagnostic test}
\usage{
diagnostic_summary(TP, FN, FP, TN, conf.level = 0.95, digits = 4)
}
\arguments{
\item{TP}{Integer. The count of results which are true positives.}

\item{FN}{Integer. The count of results which are false negatives.}

\item{FP}{Integer. The count of results which are false positives.}

\item{TN}{Integer. The count of results which are true negatives}

\item{conf.level}{Numeric, The confidence level for the confidence interval.
Defaults to 0.95.}

\item{digits}{Integer. The number of decimal places at which to round.
Defaults to 4.}
}
\value{
A data.frame with ten rows and three columns containing the point
estimate and confidence intervals for the following statistical measures:
sensitivity, specificity, positive predictive value, negative predictive
value, positive likelihood ratio, negative likelihood ratio, odds ratio,
Youden index, accuracy and error rate.
}
\description{
Calculates common measures of performance for a diagnostic or
other binary classification test, along with confidence intervals.
}
\details{
This function is adapted from the \code{diagnostic} function of the
\code{ThresholdROC} package written by Sara Perez-Jaume, Konstantina
Skaltsa, Natalia Pallares and Josep L. Carrasco (DOI:
10.18637/jss.v082.i04, see the References section for full citation).
}
\references{
Perez-Jaume S, Skaltsa K, Pallarès N, Carrasco JL (2017).
"ThresholdROC: Optimum Threshold Estimation Tools for Continuous Diagnostic
Tests in R." \emph{Journal of Statistical Software}, \emph{82}(4), 1-21. doi:
\href{https://www.jstatsoft.org/article/view/v082i04}{10.18637/jss.v082.i04}.
}
