% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{emerson_poly}
\alias{emerson_poly}
\title{Compute orthogonal polynomial contrasts using the algorithm of Emerson (1968)}
\usage{
emerson_poly(x, w = NULL, degree = length(x) - 1)
}
\arguments{
\item{x}{A numeric vector of data points.}

\item{w}{An optional vector of weights. If NULL, weights are set to a vector of ones
the length of x.}

\item{degree}{The highest order polynomial for which contrast coefficients should be
returned. Values greater than x are ignored.}
}
\value{
An array, with columns corresponding to the values of \code{x} and rows
containing the corresponding polynomial contrast coefficients for orders 0 to
\code{degree}.
}
\description{
The returns a matrix of orthogonal polynomial contrasts for any
arbitrarily spaced and weighted set of points up to the order specified in the
\code{degree} argument.
}
\examples{
# Reproduce results from page 700 of Emerson (1968):
values <- c(0, 2, 3, 5)
emerson_poly(x = values, w = c(2, 6, 6, 2))

}
\references{
Emerson, P. L. (1968). Numerical Construction of Orthogonal Polynomials
from a General Recurrence Formula. Biometrics, 24(3), 695–701.
https://doi.org/10.2307/2528328
}
