% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_comb}
\alias{formula_comb}
\title{Create formulas containing all combinations of 1 to \emph{N} predictor variables}
\usage{
formula_comb(form, n.max, marginality = T)
}
\arguments{
\item{form}{The maximal model formula.}

\item{n.max}{Integer. The largest number of predictor variables to include in the formulae.}

\item{marginality}{Logical. Should a formula be dropped if it contains an interaction
term for which one or more of the main effects is missing? Defaults to \code{TRUE}.}
}
\value{
A list of formulae
}
\description{
Create formulas containing all combinations of 1 to \emph{N} predictor variables
}
