% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcf}
\alias{gcf}
\title{Calculate the greatest common factor for a vector of values}
\usage{
gcf(x)
}
\arguments{
\item{x}{Vector of integers. NOTE that decimals will be silently dropped!}
}
\value{
Integer
}
\description{
Euclid's algorithm is employed to calculate the greatest common factor
from a set of values.
}
