% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotes_et_al.R
\name{geys}
\alias{geys}
\title{Calculate the Yield Stability Index of Cotes et al. (2002)}
\usage{
geys(
  data,
  response = "yield",
  genoytpe = "geno",
  environment = "env",
  block = NULL,
  alpha = 0.1,
  ...
)
}
\arguments{
\item{data}{data.frame. The data frame containing all the columns given by the
parameters response, environment, genotype and, optionally, block.}

\item{response}{character. The name of the column which contains the response variable.
The default is "yield". Also note that in selecting genotypes, it is assumed that
higher values of the response are better.}

\item{environment}{character. The name of the column containing the environemnt labels.
The default is "env".}

\item{block}{character. The name of the column containing the block labels. The default
is NULL, corresponding to no blocks (i.e a completely randomized design at all
locations).}

\item{alpha}{The type I error rate to use for making the LSD adjustment to the yield
rankings (see Cotes et al. (2002) for details). Defaults to 0.10.}

\item{...}{Other arguments passed to \code{nlme::lme()} during model fitting for the
estimation of Shukla's variances and SED's.}

\item{genotype}{character. The name of column containing the genotype labels. The
default is "geno".}
}
\value{
a data frame containing the yield stability rankings, along with some
intermediate values used in their calculation. \code{object}.
}
\description{
This function is ported from the SAS macro given in Cotes et al. (2002),
which those authors describe as being derived from the approach described by Kang
(1996).
}
\references{
Cotes, J. M., Ñustez, C. E., Martinez, R., & Estrada, N. (2002). Analyzing
genotype by environment interaction in potato using yield-stability index. American
Journal of Potato Research, 79(3), 211–218. https://doi.org/10.1007/BF02871937

Kang, M. S., & Magari, R. (1996). New development in selecting for
phenotypic stability in crop breeding. In M. S. Kang & H. G. Gauch (Eds.), Genotype by
Environment Interaction (First Edition). CRC-Press.
}
