% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_confint.R
\name{inverse_confint}
\alias{inverse_confint}
\title{Calculate inverse predictions with confidence intervals for probit
dose response curve}
\usage{
inverse_confint(
  n.treated,
  n.responding,
  dose,
  conf.level = 0.95,
  probs = NULL,
  ...
)
}
\arguments{
\item{n.treated}{A vector of integers describing the number of subjects
exposed at each level of dosage. The vctor must be the same length as that
supplied for \code{dose}}

\item{n.responding}{A vector of integers describing the number of exposed
subjects who responded at each level of dosage. The vector must be the
same length as that supplied for \code{dose} and no no integer may exceed
the corresponding \code{n.treated} value.}

\item{dose}{A vector of numeric values.}

\item{conf.level}{Numeric. A value between 0 and 1 controlling the width of
the confidence intervals.}

\item{...}{Other options passed to \code{glm}.}

\item{p}{An optional vector of probabilities at which to calculate dosages.
If left as \code{NULL}, dosages are calculated for probabilities ranging
from 0.01 to 0.99 at intervals of 0.01 (p < 0.1 and p > 0.9) or 0.05 (0.1
< p < 0.9)}
}
\value{
a \code{data.frame}.
}
\description{
This function takes fits a generalized linear model via the
\code{\link[stats]{glm}} function with family = binomial(link = probit),
then uses the procedures described in Finney (1971) to generate inverse
predictions and corrected confidence intervals. The code was contributed
by James Collee.
}
\references{
Finney, D.J. (1971) Probit Analysis. 3rd Edition, Cambridge
University Press: Cambridge.
}
