% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ratio.R
\name{lr.test}
\alias{lr.test}
\title{Likelihood Ratio Chi-Square Test}
\usage{
lr.test(x)
}
\arguments{
\item{x}{Array. A 2 dimensional contingency table in the form of an array or matrix.}
}
\value{
A list with class "htest" containing: the test statistic, the degrees of
freedom, and the p-value
}
\description{
Performs a likelihood ratio chi-square test.
}
