% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize}
\alias{maize}
\title{Maize (\emph{Zea mays} L.) performance under two methods of crop improvement}
\format{
A data frame containing 58 measurements:
\describe{
\item{\strong{method}:}{Breeding method, a factor with two levels "A" and "B"}
\item{\strong{hybrid}:}{Hybrid identifier, an integer}
\item{\strong{mpi}:}{Performance index, numeric value equal to 100 × (yield ÷ grain moisture)}
}
}
\source{
Peña Ortega, G. (1999). Evaluation of three recurrent selection methods in two
short-season maize (Zea mays L.) synthetics. [Doctoral dissertation, University of
Guelph]. UG Atrium Repository.
}
\usage{
maize
}
\description{
Maize (\emph{Zea mays} L.) performance under two methods of crop improvement
}
\keyword{datasets}
