% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overdisp}
\alias{overdisp}
\title{Test for overdispersion in a gitted GL(M)M}
\usage{
overdisp(model, df = NULL, digits = 2)
}
\arguments{
\item{model}{An fitted glmmTMB, glmer or glm model}

\item{df}{Numeric. The degrees of freedom for testing. Defualts to NULL, in which case
df are extracted from the fitted model using df.residual()}

\item{digits}{Integer. The number of digits to round the statistics when outputting the
the results}
}
\description{
Test for overdispersion in a gitted GL(M)M
}
\details{
The code for this function is adapted from some originally posted online
by Ben Bolker.
}
