% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lme.R
\name{part_mat}
\alias{part_mat}
\alias{inv_subV}
\title{Utilities for case deletion diagnostics for mixed models}
\usage{
part_mat(M, n, which = c(1, 2, 3, 4))
}
\arguments{
\item{M}{A (square) matrix.}

\item{n}{Integer. Denotes the row/column along which to partition the matrix M.}

\item{which}{Which of the four partitions of M matrix should be returned (see details).}

\item{V}{A marginal variance-covariance matrix.}
}
\description{
Utilities for case deletion diagnostics for mixed models
}
\details{
These functions are described in Christensen, Pearson & Johnson (1992). The
four components of the M matrix are as follows: 1 = \code{M[n, n]}, 2 = \code{M[n, -n]}, 3 =
\code{M[-n, -n]}, and 4 = \code{M[-n, n]}
}
\references{
Christensen, Ronald, Larry M. Pearson, and Wesley Johnson. 1992.
“Case-Deletion Diagnostics for Mixed Models.” Technometrics 34 (1): 38.
URL:https://doi.org/10.2307/1269550.
}
