% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcr}
\alias{pcr}
\title{Sudden Oak Death (\emph{Phytophthora ramorum} Werres et al.) detection using polymerase
chain reaction (PCR) and tissue culture}
\format{
A 2x2 array containing the results of 428 tests for Sudden Oak Death conducted
using both:
\describe{
\item{\strong{Gold Standard Test}:}{Agar culture of plant tissue}
\item{\strong{Alternate Test}:}{PCR testing for causal agent}
}
}
\source{
Vettraino, A.M., Sukno, S., Vannini, A., & Garbelotto, M. (2010). Diagnostic
sensitivity and specificity of different methods used by two laboratories for the
detection of Phytophthora ramorum on multiple natural hosts. Plant Pathology, 59,
289-300. DOI: \href{https://doi.org/10.1111/j.1365-3059.2009.02209.x}{10.1111/j.1365-3059.2009.02209.x}
}
\usage{
pcr
}
\description{
A confusion matrix showing detection/non-detection of 428 tests for
Sudden Oak Death using two diagnostic methods.
}
\keyword{datasets}
