% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pea_color}
\alias{pea_color}
\alias{pea_colour}
\title{Pea (\emph{Pisum sativum} L.) color}
\format{
A 1x2 array of integers
}
\source{
Hellens, R. P., Moreau, C., Lin-Wang, K., Schwinn, K. E., Thomson, S. J.,
Fiers, M. W. E. J., Frew, T. J., Murray, S. R., Hofer, J. M. I., Jacobs, J. M. E.,
Davies, K. M., Allan, A. C., Bendahamane, A., Coyne, C. J., Timmerman-Vaughan, G. M.,
& Ellis, T. H. N. (2010). Identification of Mendel’s white flower character. PLOS
One, 5(10): e1320. DOI: \href{https://www.doi.org/10.1371/journal.pone.0013230}{10.1371/journal.pone.0013230}.
}
\usage{
pea_color
}
\description{
An array containing the number of plants with white and purple flowers in a sample of 81 pea (\emph{Pisum sativum} L.) plants.
}
\keyword{datasets}
