% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompanion.R
\name{plot_normal_histogram}
\alias{plot_normal_histogram}
\title{Histogram with normal curve}
\usage{
plot_normal_histogram(
  x,
  by = NULL,
  bar.opts = list(),
  line.opts = list(),
  kernel = F
)
}
\arguments{
\item{x}{A vector of numeric values to plot.}

\item{by}{A vector indicating grouping variables. A separate plot is generated for
each unique value present in \code{by}.}

\item{bar.opts}{A named list of optional arguments passed to
\code{\link[graphics]{hist}}}

\item{line.opts}{A named list of optional arguments passed to
\code{\link[graphics]{curve}}}

\item{kernel}{Logical. Should a kernal density estimate be added to the plot? Defaults
to FALSE.}
}
\value{
\code{NULL}. Produces a plot.
}
\description{
Plots a histogram and superimposes on it a normal probability
density curve with mean and sigma parameters calculated from the data.
}
\details{
This function is loosely adapted from the \code{plotNormalHistogram}
function of the \code{rcompanion} package written by Salvatore Mangiafico (
\url{https://CRAN.R-project.org/package=rcompanion}, see the References
section for full citation).
}
\references{
Mangiafico, S. (2021). "rcompanion: Functions to Support
Extension Education Program Evaluation". R package version 2.3.27.
\url{https://CRAN.R-project.org/package=rcompanion}
}
