% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{post_hoc_power}
\alias{post_hoc_power}
\title{Post hoc power analysis of contrasts}
\usage{
post_hoc_power(object, alpha = 0.05, denDF = NULL)
}
\arguments{
\item{object}{An emmGrid object (from the emmeans package): the result of a call
contrast(emmeans(model))}

\item{alpha}{Numeric. Coverage for confidence intervals.}
}
\value{
A data frame containing the power, along with values used to calculate it, for
each contrast.
}
\description{
This function calculates power of contrasts specified in an emmGrid
object. Code is adapted from the SAS macro powercalc, by Stephen Bowley.
}
