% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcomp.R
\name{predict_cs_conditional}
\alias{predict_cs_conditional}
\title{Calculate the random effects estimates implied in a marginal (compound symmetric)
model}
\usage{
predict_cs_conditional(object, data)
}
\arguments{
\item{object}{An \code{\link[nlme]{lme}} or \code{\link[nlme]{gls}} object containing a
corCompSymm correlation structure.}

\item{data}{The data frame used in fitting \code{object}. Required if \code{object} is
of class \code{gls}}
}
\value{
A vector of random effects estimates.
}
\description{
Given the mathematical equivalence between a conditional model with independent
random effects and a compound symmetric correlation structure, its possible to calculate the
random effects estimates implied by the latter:
\deqn{\mathbf{G}\mathbf{Z}'\mathbf{V}^{-1}(Y - \mathbf{X}\beta)}{GZ'V^ (-1) (Y - X \beta)}
}
\details{

}
\references{
Stroup, W., Milliken, G., Claasen, E., Wolfinger, R. (2018).
"SAS for Mixed Models: Introduction and Basic Applications". Cary, NC: SAS
Institute, Inc.
}
