% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_condensed}
\alias{print_condensed}
\title{Print compact versions of data.frames, matrices and arrays}
\usage{
print_condensed(x, rows = NULL, cols = NULL, digits = 0, warn = F)
}
\arguments{
\item{x}{A data.frame, matrix or array, or any object which can be coerced to a
data.frame.}

\item{rows}{A vector of two integers giving the number of initial and final rows of
\code{x} to include in the output, respectively. If only a single value is provided,
that number is used for the number of both initial and final rows of \code{x} to
include. Defaults to NULL, in which case all rows are included.}

\item{cols}{A vector of two integers giving the number of initial and final columns of
\code{x} to include in the output, respectively. If only a single value is provided,
that number is used for both the number of initial and final columns of \code{x} to
include. Defaults to \code{NULL}, in which case all columns are included.}

\item{digits}{The number of digits at which to round numeric columns in \code{x}}

\item{warn}{Logical. Should a warning be provided if invalid arguments are provided for
\code{rows} and/or \code{cols}? Defaults to \code{FALSE}.}
}
\value{
Creates and prints a \code{data.frame} of all character vectors, where the
middle rows and/or columns (defined by the arguments \code{rows} and \code{cols})
have been omitted, and replaced with a single row and/or column of ellipses ('...').
}
\description{
This is a convenience function for concisely and compactly displaying
large data frames, matrices and arrays using ellipses to indicate omitted rows and/or
columns. All columns in the outputted \code{data.frame}, \code{matrix}, or
\code{array} are character values, and thus the output of this function should only be
used for printing or displaying data sets.
}
\details{
If \code{x} is a data.frame, original column and row names are retained and
printed. If it is an array with non-\code{NULL} dimnames, new column names are
created (and printed) from them. If \code{x} is a matrix, the output is printed
without row or column names. If invalid values are provided for \code{rows} and/or
\code{cols} arguments, the invalid value(s) are ignored (silently if \code{warn = F})
and all rows and/or columns are printed.
}
