% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcomp.R
\name{eblup_terms}
\alias{eblup_terms}
\title{Extract the fixed and random effects in a fitted model, ordered for use in
specifying estimable linear functions.}
\usage{
eblup_terms(object)
}
\arguments{
\item{object}{A model fitted with \code{\link[nlme]{lme}} or \code{\link[lme4]{lmer}}.}
}
\value{
A column vector of the names of fixed- and random-effects coefficients in \code{object}.
}
\description{
This function provides the names of the terms, in proper order, for
specifying contrast statements in calls to eblup(). Note that to correctly calculate
the standard errors of eblups, the ordering of the terms must agree with that of the
projection matrix, which may differ from the ordering of terms provided by calls to
fixef() and/or ranef().
}
\examples{
library(HGSB)
\dontrun{library(lme4)}
data(trefoil)
fit <- lmer(seed_g ~ 1 + (1|family) + (1|block), trefoil)

eblup_terms(fit)

\dontrun{library(nlme)}
fit <- lme(seed_g ~ 1, trefoil, random = list(family = ~ 1, family = pdIdent(form = ~ block)))

eblup_terms(fit)

}
