% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcomp.R
\name{get_vcomp}
\alias{get_vcomp}
\title{Extract the \strong{Z}, \strong{G}, \strong{R} and \strong{V}
matrices from a fitted model}
\usage{
get_vcomp(x, ...)
}
\arguments{
\item{x}{A fitted model. Currently, models of the following class are supported: "lme", "merMod".}

\item{data}{The data frame to be used in constructing the design matrix. Uses the data stored in the
model object if not supplied. See details for more information.}
}
\value{
A list of matrices.
}
\description{
The marginal variance-covariance matrix \strong{V} for fitted model
mixed effects model is equal to:
V = ZGZ' + R
Where G is the random effects variance-covariance matrix, R is residual variance-covariance
matrix and Z is the random effects design matrix. These matrices are useful for calculating,
for example, EBLUPs and their standard errors.
}
\details{
This function returns the incidence matrix,
\strong{Z}, the random effects covariance
matrix, \strong{G}, the residual covariance matrix, \strong{R}, and the marginal covariance matrix, \strong{V}. These
matrices are useful for calculating the covariance matrix of the fixed and
random effects "estimates" together, which is needed for the calculation
of the square root of the estimated projection variance for EBLUPs. The
marginal variance covariance matrix \strong{V} is also returned.

For models of class "lme", the code employed is adapted from the mcgv::extract.lme.cov() function
of Wood (2021), with the most significant modification being adjustments which correct
the G and Z matrices when crossed random effects are specified.
}
\references{
Stroup, W., Milliken, G., Claasen, E., Wolfinger, R. (2018).
"SAS for Mixed Models: Introduction and Basic Applications". Cary, NC: SAS
Institute, Inc.

Wood, S. (2021). mgcv: Mixed GAM Computation Vehicle with
Automatic Smoothness Estimation (1.8-36). \code{\url{https://CRAN.R-project.org/package=mgcv}}
}
